+++ apps/app/drizzle/0014_aspiring_namor.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "node_versions" ADD COLUMN "created" boolean DEFAULT false NOT NULL;